centralize managing links
