+++ src/objects/ErrorObject.php
@@ -5,13 +5,13 @@ namespace alsvanzelf\jsonapi\objects;
 use alsvanzelf\jsonapi\Document;
 use alsvanzelf\jsonapi\exceptions\InputException;
 use alsvanzelf\jsonapi\helpers\Converter;
+use alsvanzelf\jsonapi\helpers\LinksManager;
 use alsvanzelf\jsonapi\helpers\ManageHttpStatusCode;
 use alsvanzelf\jsonapi\helpers\Validator;
 use alsvanzelf\jsonapi\interfaces\ObjectInterface;
-use alsvanzelf\jsonapi\objects\LinksObject;
 
 class ErrorObject implements ObjectInterface {
-	use ManageHttpStatusCode;
+	use LinksManager, ManageHttpStatusCode;
 	
 	/** @var string */
 	protected $id;
@@ -21,8 +21,6 @@ class ErrorObject implements ObjectInterface {
 	protected $title;
 	/** @var string */
 	protected $detail;
-	/** @var LinksObject */
-	protected $links;
 	/** @var array */
 	protected $source = [];
 	/** @var MetaObject */
@@ -130,19 +128,6 @@ class ErrorObject implements ObjectInterface {
 	}
 	
 	/**
-	 * @param string $key
-	 * @param string $href
-	 * @param array  $meta optional, if given a LinkObject is added, otherwise a link string is added
-	 */
-	public function addLink($key, $href, array $meta=[]) {
-		if ($this->links === null) {
-			$this->setLinksObject(new LinksObject());
-		}
-		
-		$this->links->add($key, $href, $meta);
-	}
-	
-	/**
 	 * set the link about this specific occurence of the error, explained in a human-friendly way
 	 * 
 	 * @param string $href
@@ -235,13 +220,6 @@ class ErrorObject implements ObjectInterface {
 	 */
 	public function setHumanDetails($specificDetails) {
 		$this->detail = $specificDetails;
-	}
-	
-	/**
-	 * @param LinksObject $linksObject
-	 */
-	public function setLinksObject(LinksObject $linksObject) {
-		$this->links = $linksObject;
 	}
 	
 	/**
